+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(theme): enhance high-contrast theme with updated colors and styles. (`a1c80a5`)
 - chore(dependencies): update @harborclient/sdk to version 1.1.22 and enhance theme configurations. (`4685d66`)
 - chore(catalog): update theme references to latest versions. (`89f7852`)